+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Create a hook for the controller to mutate search values (e.g. for date formatting) (https://hitower.atlassian.net/browse/WEAP-187).
 - Print any invalid relation names in the exception.
 - Laravel Pint integrated.
+- Added the Temporary Media Upload functionality.
 
 ## v0.0.15
 - Return only the pagination keys if the request is paginated.